chore :: 함수 분리
