+++ src/main/kotlin/dsm/wemeet/global/mail/presentation/MailController.kt
@@ -2,7 +2,8 @@ package dsm.wemeet.global.mail.presentation
 
 import dsm.wemeet.global.mail.presentation.dto.request.CheckMailRequest
 import dsm.wemeet.global.mail.presentation.dto.request.SendMailRequest
-import dsm.wemeet.global.mail.service.MailService
+import dsm.wemeet.global.mail.service.SendMailService
+import dsm.wemeet.global.mail.service.VerifyMailService
 import jakarta.validation.Valid
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -12,18 +13,19 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/mail")
 class MailController(
-    private val mailService: MailService
+    private val sendMailService: SendMailService,
+    private val verifyMailService: VerifyMailService
 ) {
 
     @PostMapping
     fun sendMail(
         @Valid @RequestBody
         request: SendMailRequest
-    ) = mailService.sendCode(request)
+    ) = sendMailService.execute(request)
 
     @PostMapping("/check")
     fun verify(
         @Valid @RequestBody
         request: CheckMailRequest
-    ) = mailService.verifyMailCode(request)
+    ) = verifyMailService.execute(request)
 }
