+++ src/main/kotlin/dsm/wemeet/global/mail/service/SendMailService.kt
@@ -1,0 +1,58 @@
+package dsm.wemeet.global.mail.service
+
+import dsm.wemeet.domain.user.service.QueryUserService
+import dsm.wemeet.global.mail.MailProperties
+import dsm.wemeet.global.mail.exception.InternationalMailServerException
+import dsm.wemeet.global.mail.presentation.dto.request.SendMailRequest
+import dsm.wemeet.global.redis.RedisUtil
+import jakarta.mail.internet.InternetAddress
+import jakarta.mail.internet.MimeMessage
+import org.springframework.core.io.ClassPathResource
+import org.springframework.mail.javamail.JavaMailSender
+import org.springframework.mail.javamail.MimeMessageHelper
+import org.springframework.stereotype.Service
+import org.springframework.util.FileCopyUtils
+import kotlin.random.Random
+
+@Service
+class SendMailService(
+    private val mailSender: JavaMailSender,
+    private val mailProperties: MailProperties,
+    private val queryUserService: QueryUserService,
+    private val redisUtil: RedisUtil
+) {
+
+    fun execute(request: SendMailRequest) {
+        val cleanEmail = request.mail.replace("\"", "")
+        queryUserService.existsByEmail(cleanEmail)
+
+        val authCode = generateAuthCode()
+        redisUtil.setDataExpire(cleanEmail, authCode, 600)
+
+        val message: MimeMessage = mailSender.createMimeMessage()
+        val helper = MimeMessageHelper(message, true, "UTF-8")
+
+        helper.setTo(cleanEmail)
+        helper.setSubject("[wemeet] 이메일 인증번호 확인")
+        helper.setFrom(InternetAddress(mailProperties.username, "wemeet"))
+
+        val content = loadEmailTemplate().replace("\${authCode}", authCode)
+
+        helper.setText(content, true)
+        try {
+            mailSender.send(message)
+        } catch (e: Exception) {
+            redisUtil.deleteData(cleanEmail)
+            throw InternationalMailServerException
+        }
+    }
+
+    private fun generateAuthCode(): String {
+        return Random.nextInt(1000, 9999).toString()
+    }
+
+    private fun loadEmailTemplate(): String {
+        val template = ClassPathResource("templates/email_template.html")
+        return String(FileCopyUtils.copyToByteArray(template.inputStream))
+    }
+}
